+++ CHANGELOG.md
@@ -14,8 +14,6 @@ The **patch** part changes is incremented if multiple releases happen the same m
 
 ### Changed
 
-* minifirewall: tail template follows symlinks
-
 ### Fixed
 
 ### Removed
@@ -36,6 +34,7 @@ The **patch** part changes is incremented if multiple releases happen the same m
 
 * docker: Allow "live-restore" to be toggled with docker_conf_live_restore
 * evocheck: upstream release 22.06
+* minifirewall: tail template follows symlinks
 * mysql: add "set crypt_use_gpgme=no" Mutt option, for mysqltuner
 
 ## [22.05.1] 2022-05-12