+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The **patch** part changes is incremented if multiple releases happen the same m
 
 ### Changed
 
+* mysql: reorganize evomariabackup to use mtree instead of our own dir-check
+
 ### Fixed
 
 ### Removed